review: split integration version tests, wrap lock errors, refresh comments (#274)
Round 5 mechanical items, no behavior change:
- postgres_persistent_repository_integration_test.go (539) splits the
  three #274 version-contract anchors into
  postgres_persistent_repository_version_integration_test.go
  (288 + 281), and the concurrency sweep moves into a named helper,
  bringing TestConcurrentDeleteRecreateSerializesIntegration to 79
  lines.
- The single-row lockRowVersion returns wrap with their operation,
  matching the batch paths.
- pausing.go no longer describes the retired changed_at <= snapshot
  predicate (the mark guard is exact listed-version equality);
  merge_sql.go's buildMergeSQL doc drops the retired base-wins framing
  for the #274 canonical-0 one.
- Ride-alongs from the review's observations: chooseLWW notes the
  Go-vs-SQL evaluation-order nuance (semantics match, order differs,
  divergence unreachable post-#274, noted for #365), and design.md's
  live-encoding sentence names the benchmark harness shape as the
  deliberate raw-encoding holdout.
